feat: add pluggable eslint rules for imports
Reasoning to add those is to be able to entirely remove on built in
functionality of vs code that currently allows us to organise imports.
Switching to eslint for that removes potential problems with race
conditions and double formatting.
In order to use new rules add extension to your eslint configuration.
You can find examples on how to do that in sample app.